+++ app/Http/Requests/MachineRequest.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class MachineRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'site_id' => 'required',
+            'name' => 'required|max:255',
+            'ip' => 'required|ipv4',
+        ];
+    }
+
+    public function attributes()
+    {
+        return [
+            'site_id' => 'Standort',
+            'name' => 'Name',
+            'ip' => 'IP',
+        ];
+    }
+}