+++ resources/views/components/create/select/recorder.blade.php
@@ -1,0 +1,8 @@
+<div class="flex flex-col mt-2">
+    <x-input.label for="recorder_id" value="Recorder" />
+    <x-input.select id="recorder_id" name="recorder_id">
+        @foreach ($recorders as $recorder)
+            <option value="{{ $recorder->id }}">{{ $recorder->name }}</option>
+        @endforeach
+    </x-input.select>
+</div>